refactor: extract AppState session-library extension (#576, #430)
Fourth slice of #430 (AppState decomposition). Second non-adjacent
slice after #574 (window-management). Extracts 4 session-library
delegator methods from two non-adjacent clusters into a new
Sources/BugNarrator/AppState+SessionLibrary.swift as `extension
AppState { ... }`.

**Cluster A** (AppState.swift lines 969-971, 3 lines, 160 bytes, SHA-256
68e076d0811566821f0aaab35c0036c44e160bcea8657e477dfdd7a1d92ce843):
copyDisplayedTranscript() — simple presenter delegation.

**Cluster B** (AppState.swift lines 1038-1060, 23 lines, 804 bytes,
SHA-256 755f9923ef1f54361c1f41e3304c08851723b3a07fa3816ee01b0b06faaa3616):
saveCurrentTranscriptToHistory, deleteDisplayedTranscript,
deleteSessions(withIDs:) — all three use identical do-catch shape
(try sessionLibrary.<method>() + presenter.present<successName> +
catch → presenter.presentFailure(error)).

**retryPendingTranscription(for:)** (async orchestrator, starts at line
973 pre-move) has real behavior (retry-context logic) and STAYS in
AppState.swift.

Applied the full protocol stack: spec-review-verify-SHA (both clusters
matched origin at spec review), endpoint verification (969 = `func
copyDisplayedTranscript() {`, 971 = `}`, 972 = blank, 973 =
retryPendingTranscription — the async orchestrator that stays; 1038 =
`func saveCurrentTranscriptToHistory() {`, 1060 = `}` of deleteSessions,
1061 = blank), post-extract per-range byte-diff against origin (both
silent), reverse-order sed deletion per #556 non-adjacent protocol
(1038-1061 first, then 969-972).

No dead-import cleanup needed — codex confirmed AppState.swift still
needs Combine (ObservableObject / @Published) and Foundation (UUID,
URL, TimeInterval). New file needs only `import Foundation` (Set<UUID>
uses UUID).

The extension does NOT re-annotate @MainActor — inherited from
AppState's class-level annotation per SE-0316 (same as prior 430 slices).
662 unit tests pass unchanged; docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (2 findings folded:
line 972 is trailing blank not retryPendingTranscription start, xcodegen
project-file update — both applied).
